refactor(grid): make grid-item a separate directory
