+++ src/action-sheet/action-sheet.json
@@ -4,7 +4,7 @@
     "t-icon": "../icon/icon",
     "t-popup": "../popup/popup",
     "t-grid": "../grid/grid",
-    "t-grid-item": "../grid/grid-item",
+    "t-grid-item": "../grid-item/grid-item",
     "t-swiper-nav": "../swiper-nav/swiper-nav"
   }
 }
