+++ src/grid-item/grid-item.ts
@@ -1,6 +1,6 @@
 import { SuperComponent, wxComponent, isObject, RelationsOptions } from '../common/src/index';
 import config from '../common/config';
-import props from './grid-item-props';
+import props from './props';
 import { uniqueFactory, setIcon } from '../common/utils';
 
 const { prefix } = config;
@@ -30,7 +30,7 @@ export default class GridItem extends SuperComponent {
   };
 
   relations: RelationsOptions = {
-    './grid': {
+    '../grid/grid': {
       type: 'ancestor',
       linked(target) {
         this.parent = target;
