+++ src/grid/README.en-US.md
@@ -18,11 +18,11 @@ theme | String | default | options：default/card | N
 
 name | type | default | description | required
 -- | -- | -- | -- | --
-badge-props | Object | null | Typescript：`BadgeProps`，[Badge API Documents](./badge?tab=api)。[see more ts definition](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/grid/type.ts) | N
+badge-props | Object | null | Typescript：`BadgeProps`，[Badge API Documents](./badge?tab=api)。[see more ts definition](https://github.com/Tencent/tdesign-miniprogram/tree/develop/src/grid-item/type.ts) | N
 custom-style | String | - | \- | N
 description | String / Slot | - | \- | N
 external-classes | Array | - | `['t-class', 't-class-image', 't-class-text', 't-class-description']` | N
-icon | String | - | \- | N
+icon | String / Object | - | \- | N
 image | String / Slot | - | \- | N
 image-props | Object | - | \- | N
 jump-type | String | navigate-to | options：redirect-to/switch-tab/relaunch/navigate-to | N
