+++ src/grid/_example/badge/index.json
@@ -2,6 +2,6 @@
   "component": true,
   "usingComponents": {
     "t-grid": "tdesign-miniprogram/grid/grid",
-    "t-grid-item": "tdesign-miniprogram/grid/grid-item"
+    "t-grid-item": "tdesign-miniprogram/grid-item/grid-item"
   }
 }
