+++ src/grid/grid.ts
@@ -10,7 +10,7 @@ export default class Grid extends SuperComponent {
   externalClasses = ['t-class'];
 
   relations: RelationsOptions = {
-    './grid-item': {
+    '../grid-item/grid-item': {
       type: 'descendant',
     },
   };
