+++ src/upload/upload.json
@@ -2,7 +2,7 @@
   "component": true,
   "usingComponents": {
     "t-grid": "../grid/grid",
-    "t-grid-item": "../grid/grid-item",
+    "t-grid-item": "../grid-item/grid-item",
     "t-popup": "../popup/popup",
     "t-icon": "../icon/icon",
     "t-image": "../image/image"
